+++ src/Handler/StreamHandler.php
@@ -1,0 +1,82 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Yiisoft\VarDumper\Handler;
+
+use RuntimeException;
+use Yiisoft\VarDumper\HandlerInterface;
+
+final class StreamHandler implements HandlerInterface
+{
+    /**
+     * @var callable
+     */
+    private $encoder;
+    /**
+     * @var null|resource
+     */
+    private $stream = null;
+
+    public function __construct(
+        private $uri = 'udp://127.0.0.1:8890'
+    ) {
+        if (!is_string($this->uri) && !is_resource($this->uri)) {
+            throw new \InvalidArgumentException(
+                sprintf(
+                    'Argument $uri must be a string or a resource, "%s" given.',
+                    gettype($this->uri)
+                )
+            );
+        }
+    }
+
+    /**
+     * Sends encode with {@see \json_encode()} function $variable to a UDP socket.
+     */
+    public function handle(mixed $variable, int $depth, bool $highlight = false): void
+    {
+        $data = ($this->encoder ?? 'json_encode')($variable);
+        if (!is_string($data)) {
+            throw new RuntimeException(
+                sprintf(
+                    'Encoder must return a string, %s returned.',
+                    gettype($data)
+                )
+            );
+        }
+
+        $this->initializeStream();
+
+        if (!$this->writeToStream($data)) {
+            $this->initializeStream();
+
+            $this->writeToStream($data);
+        }
+    }
+
+    public function withEncoder(callable $encoder): HandlerInterface
+    {
+        $new = clone $this;
+        $new->encoder = $encoder;
+        return $new;
+    }
+
+    private function initializeStream(): void
+    {
+        if (is_string($this->uri)) {
+            $this->stream = fsockopen($this->uri);
+        } else {
+            $this->stream = $this->uri;
+        }
+
+        if (!is_resource($this->stream)) {
+            throw new RuntimeException('Cannot initialize a stream.');
+        }
+    }
+
+    private function writeToStream(bool|string $data): bool
+    {
+        return @fwrite($this->stream, $data) !== false;
+    }
+}