+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## 1.7.0 under development
 
-- New #95: Add `UdpHandler` and `CompositeHandler` (@xepozz)
+- New #95: Add `StreamHandler` and `CompositeHandler` (@xepozz)
 - Enh #94: Add a middleware handler to control dumps' output destination (@xepozz)
 - New #94: Add `dump` function (@xepozz)